Resolve conflicts with #26160 in numpy/lib/_stride_tricks_impl.py and
numpy/lib/tests/test_stride_tricks.py in favour of the nditer-based
implementation.

+++ numpy/linalg/lapack_lite/f2c.c
@@ -191,7 +191,7 @@ integer i_dnnt(x) doublereal *x;
 integer i_dnnt(doublereal *x)
 #endif
 {
-return( (*x)>=0 ?
+return (integer)( (*x)>=0 ?
 	floor(*x + .5) : -floor(.5 - *x) );
 }
